Added handle of todo item
