+++ src/redux/store.js
@@ -1,12 +1,12 @@
 import { configureStore } from "@reduxjs/toolkit";
 import todoReducer from "./todoSlice";
-import styleReducer from "./styleSlice";
+import commonReducer from "./commonSlice";
 import listReducer from "./listSlice";
 
 export default configureStore({
   reducer: {
     todos: todoReducer,
     lists: listReducer,
-    style: styleReducer,
+    style: commonReducer,
   },
 });
